refactor: move api call to core
