+++ core/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
+    id 'kotlinx-serialization'
 }
 
 apply from: "$rootDir/dependencies.gradle"
@@ -15,6 +16,7 @@ android {
         versionCode versions.versionCode
         versionName versions.versionName
         consumerProguardFiles "consumer-rules.pro"
+        buildConfigField "String", "VERSION_NAME", "\"$versions.versionName\""
     }
 
     buildTypes {
@@ -39,9 +41,12 @@ android {
 dependencies {
     api project(':wallet')
     implementation "androidx.core:core-ktx:$versions.androidx_core"
+    implementation "androidx.browser:browser:$versions.androidx_browser"
     implementation "androidx.appcompat:appcompat:$versions.androidx_appcompat"
     implementation "com.google.android.material:material:$versions.material"
-    implementation "androidx.browser:browser:$versions.androidx_browser"
+    implementation "com.squareup.okhttp3:okhttp:$versions.okhttp"
+    implementation "com.squareup.okhttp3:logging-interceptor:$versions.okhttp"
+    implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$versions.serialization"
 
     testImplementation "junit:junit:$versions.junit"
     testImplementation "io.mockk:mockk:$versions.mockk"
