+++ core/src/main/java/com/portto/sdk/core/network/service/FlowService.kt
@@ -1,0 +1,14 @@
+package com.portto.sdk.core.network.service
+
+import androidx.annotation.WorkerThread
+import com.portto.sdk.core.network.BloctoApi
+import kotlinx.serialization.Serializable
+
+
+object FlowService {
+    @WorkerThread
+    fun getFeePayer(): FeePayerResponse = BloctoApi.get("flow/feePayer")
+
+    @Serializable
+    data class FeePayerResponse(val address: String)
+}
