+++ flow/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlinx-serialization'
 }
 
 apply from: "$rootDir/dependencies.gradle"
@@ -36,9 +35,7 @@ android {
 dependencies {
     api project(':core')
     implementation "androidx.core:core-ktx:$versions.androidx_core"
-    implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$versions.serialization"
-    implementation "com.squareup.okhttp3:okhttp:$versions.okhttp"
-    implementation "com.squareup.okhttp3:logging-interceptor:$versions.okhttp"
+
     testImplementation "junit:junit:$versions.junit"
     testImplementation "org.robolectric:robolectric:$versions.robolectric"
 }
