+++ solana/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlinx-serialization'
 }
 
 apply from: "$rootDir/dependencies.gradle"
@@ -16,7 +15,6 @@ android {
         versionCode versions.versionCode
         versionName versions.versionName
         consumerProguardFiles "consumer-rules.pro"
-        buildConfigField "String", "VERSION_NAME", "\"$versions.versionName\""
     }
 
     buildTypes {
@@ -38,9 +36,6 @@ dependencies {
     api project(':core')
     implementation files("libs/kbase58.jar")
     implementation "androidx.core:core-ktx:$versions.androidx_core"
-    implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$versions.serialization"
-    implementation "com.squareup.okhttp3:okhttp:$versions.okhttp"
-    implementation "com.squareup.okhttp3:logging-interceptor:$versions.okhttp"
     implementation "com.portto.solana:web3:$versions.solana_web3"
 
     testImplementation "junit:junit:$versions.junit"
